move animations to utils
